@@ -1,61 +1,33 @@
 package ru.slavapmk.shtp.ui;
 
+import static ru.slavapmk.shtp.ui.MainActivity.fmanager;
+
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageButton;
+
 import androidx.fragment.app.Fragment;
-import static ru.slavapmk.shtp.ui.MainActivity.fmanager;
 
 import ru.slavapmk.shtp.R;
 
 public class ProfileFragment extends Fragment {
-    private String mParam1;
-    private String mParam2;
-
     @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View inflate = inflater.inflate(R.layout.fragment_profile, container, false);
-
-        ImageButton imageButton = inflate.findViewById(R.id.imageButton11);
-        Log.d("TAG", imageButton.toString());
-        imageButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                fmanager.beginTransaction()
+        inflate.findViewById(R.id.imageButton11).setOnClickListener(view ->
+                fmanager
+                        .beginTransaction()
                         .setReorderingAllowed(true)
                         .remove(ProfileFragment.this)
-                        .add(R.id.fragmentContainer, SettingsFragment.newInstance("", ""))
+                        .add(R.id.fragmentContainer, SettingsFragment.newInstance())
                         .addToBackStack("backstack")
-                        .commit();
-            }
-        });
-
+                        .commit()
+        );
         return inflate;
     }
 
-    public static final String ARG_PARAM1 = "param1";
-    public static final String ARG_PARAM2 = "param2";
-
-    public static ProfileFragment newInstance(String param1, String param2) {
-        ProfileFragment fragment = new ProfileFragment();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
+    public static ProfileFragment newInstance() {
+        return new ProfileFragment();
     }
 }
